Support configuring the `format` in `st.metric`
