+++ e2e_playwright/st_metric.py
@@ -138,3 +138,18 @@ st.metric("Pixel height (200px)", 123, 123, border=True, height=200)
 with st.container(height=400, key="height_test"):
     st.metric("Stretch height", 123, 123, height="stretch")
     st.metric("Content height", 123, 123, height="content")
+
+# Format parameter tests
+format_col1, format_col2, format_col3, format_col4 = st.container(
+    key="metric_format_config"
+).columns(4)
+
+with format_col1:
+    st.metric("Compact format", 1234567, delta=50000, format="compact")
+with format_col2:
+    st.metric("Dollar format", 1234.56, delta=-50.25, format="dollar")
+with format_col3:
+    st.metric("Percent format", 0.1234, delta=0.05, format="percent")
+with format_col4:
+    # Non-numeric string should NOT be formatted
+    st.metric("Non-numeric (no format)", "70 °F", delta="+5%", format="compact")
